exec: Provide a way to cleanup the temporary dir
ioutil.TempDir passes the responsibility of cleanup onto us. We weren't
doing this, which leads to a temp dir that is full of runfile- temp
entries.

Thus, introduce a way to cleanup after ourselves.
This isn't exactly a problem, as most modern distributions
automatically clear stale entries out of /tmp, but it looks rather messy.